Renaming iec-60908b-math to iec-60908b
